+++ SF2DWrapper/src/SF2DWrapper.cpp
@@ -348,12 +348,12 @@ void SFSimulator::setAgentTimeHorizonObst(int agentNo, float timeHorizonObst)
 	_sim->setAgentTimeHorizonObst(agentNo,timeHorizonObst) ;
 }
 
-float SFSimulator::getAgentPressure(int index)
+double SFSimulator::getAgentPressure(int index)
 {
 	return _sim->getAgentPressure(index);
 }
 
-float SFSimulator::getObstaclePressure(int index)
+double SFSimulator::getObstaclePressure(int index)
 {
 	return _sim->getObstaclePressure(index);
 }
